refactor: 액티비티 패키지 변경
