+++ app/src/main/java/woowacourse/movie/presentation/adapter/SpinnerAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation
+package woowacourse.movie.presentation.adapter
 
 import android.content.Context
 import android.widget.ArrayAdapter
